requirements.txt
scipy
numpy
matplotlib
tqdm
seaborn
statsmodels.api